rename actions/index to actions/decks, update imports
