+++ .expo/packager-info.json
@@ -2,8 +2,8 @@
   "devToolsPort": 19002,
   "expoServerPort": 19000,
   "packagerPort": 19001,
-  "packagerPid": 23852,
+  "packagerPid": 19400,
   "expoServerNgrokUrl": "https://di-ta4.anonymous.udacicards.exp.direct",
   "packagerNgrokUrl": "https://packager.di-ta4.anonymous.udacicards.exp.direct",
-  "ngrokPid": 3696
+  "ngrokPid": 25272
 }
