+++ reducers/index.js
@@ -1,4 +1,4 @@
-import { RECEIVE_DECKS, ADD_DECK } from "../actions";
+import { RECEIVE_DECKS, ADD_DECK } from "../actions/decks";
 
 function decks(state = {}, action) {
   switch (action.type) {
